Added default program run on startup and simulation of this feature in m8rsim
Added Application class which tries to open a file called 'main' which contains
the name of the program to run (without suffix). If found it tries <name>.m8rb
and if it doesn't exit it tries <name>.m8r. If neither exit it tries main.m8rb
then main.m8r. If neither of those exist it returns with a null program. Added
the ability to ExecutionUnit to run a null program.