Fix system vs user includes to deal with the system file gpio.h vs the user file GPIO.h
The SDK contains a file called gpio.h. m8rscript has a local class file called GPIO.h. Since we're on a non-case sensitive filesystem, these names collide. This is fine as long as one is considered a "system" file and the other is a "user" file. You can use quotes to include the user file and brackets to include the system file. But to do that we need to make sure all the includes in the makefile are segregated properly and all the includes correctly use quotes or brackets. This does that.