Fixed some crashes in ESP.
Trying to put an array of C++ structures in ROM crashes, most likely because it needs to run global constructors at runtime. So I made the _pins array into a simple uin8t_t array and then combine the bytes into a PinEntry structure on copy. Esp still doesn't run the heartbeat, but it doesn't crash now. I think there's an issue with running tasks on the Esp.