Got rid of string pointer in Value
Program now has a vector of String objects that are allocated with a StringId returned, similar to Objects. This StringId is stored in the Value rather than a bare pointer. StringLiterals are now fully supported in Value. For the most part Type::String and Type::StringLiteral are identical except in the way the actual string is extracted from the Id.