Switched to 64 bit Float
In preparation for making Value a bare 64 bit value. Float will ultimately have its lowest 4 bits overloaded with a type enum. When Float sees the value, these 4 bits will be cleared. This reduces precision a bit, but we still have 8 decimal digits, which is plenty.

Had to fix several bugs with integral storage sizes to get everything to work