GC optimization for Array
Array now tracks whether or not it has any elements that need GC. Value now has a needsGC() method which returns true for Objects or Strings. Array has a _needsGC flag which is set if any NeedsGC() values are added. If not set during GC, array is not traversed. To keep the value fresh, _needsGC is recomputed during GC traversal