Implemented switch statement
Syntactically mostly like JavaScript. case values are expressions. Current implementation compares switch expression result with each case expression in turn and executes the statement when a match is found. When a match is found and the statement is executed, switch finishes. It doesn’t try to find more matches. Not sure if that is semantically like JS or not. The other big difference is that cases don’t fall through, so no break is necessary. This makes the implementation a bit more straightforward and I have never, ever seen a compelling reason for case statements to fall through.