Fixed bug with 2 cases on the same statement and optimized generated code
The previous implementation didn’t work with cases like:

    case 1:
    case 2: /* do something */

So I fixed that and in doing so I made the code much more efficient. It used to generate the sequence EQ …, JF…, <stmt>, JMP …, and so on. Now I have all the tests up from with a EQ …, JT… sequence and a JMP at the end which either goes to the default statement or the end of the switch.