Fixed a bug in CodePrinter where it was not printing nested functions properly
I made the initialization of Id template values with raw values explicit to avoid subtle errors in referencing properties, which is what was causing the bug. Now both named and anonymous functions are being output. There was another bug where anonymous functions were not generating the right code, so I fixed that, too.