Got rid of Date, Serial and System objects
For now, these objects don’t really do much, so I got rid of them and moved the 4 functions that were in them directly into Global. I also renamed Date.now() to currentTime() for clarity. Really the “Serial” object is a misnomer. It actually sends prints to WiFi so the begin() method, which set the baud rate was useless.