Got rid of several more virtual methods
Virtual methods add 4 bytes to each Object subclass and there are currently 13 subclasses. So getting rid of one saves a lot of RAM. I got rid of addProperty and addElement and absorbed their functionality into setProperty and setElement. I got rid of propertyName and propertyCount and absorbed their functionality into the iteration functionality. Finally I collapsed iteratorCount() and iterator() into a single iterate() method. Passing -1 to it returns the count as a Value, otherwise it returns the iterated value at that index.