Got Iterator::next() working
The problem with passing a "this" pointer to a call function is that it resolves to a generic object, which makes it hard to use in a native function. You'd need to verify that it's the right type and then cast it. So I added  CALLPROP opcode. When a call is being emitted by the parser, it looks to see if the callee is a PropRef or EltRef. If so, rather than dereffing, it generates a new callProperty() virtual method. This calls the property on the object. Implementing this (for instance) in the Iterator object means that the next method is called with a class method so the entire object is available.

I may eventually get rid of all CALL ops and just use CallProperty everywhere. But that's for later