Completed initial implementation of event handling.
When the program is finished, we now wait for events to come in. ExecutionManager is an EventListener and has a queue of Values. Every 50ms ExecutionManager wakes up and calls all the values in the queue. TCPSocket adds an event when its delegate fires and execution manager runs the function passed to the TCPSocket ctor when a socket is opened.

None of this is actually functional yet. There is a bug in calling the function and params are not yet passed. But the function does get called