Getting proto and passing this pointer working
2 major changes here:

1) Properties now look at prototype chain to find values
2) this is fully supported. It's passed to property calls, saved on the stack and searched during derefId().

Also made "null" property sort of work. There is an empty value at constant location 0, but it wasn't being referenced correct. So I fixed that. And now I am using that as the value when null is assigned to something. But this doesn't work correctly because you can't distinguish between assigning null and a failed assignment. So that still needs to be fixed.

Also made PropertyObject::construct first try calling it __construct value and if that's not present it calls Object::construct which now constructs a new MaterObject and sets itself as the prototype.