Fixed bug in parser generation when multiple values on parse stack need baking
Before it would pop the register after baking, which would cause the next bake to reuse that register, resulting in completely incorrect code. Now the value is left on the stack so it stays “allocated”. A new swap() function is used to position the second stack entry for baking, then it’s swapped back so they are in their original order.