Simplified Scanner keywords to make it easier to add new ones
Changed Scanner keywords to be a single string with each keyword prefixed by its token byte. Wrote a ROMstrstr() function to search this string for the desired keyword. When found, the token in the previous byte is cast and returned. To do this I had to restructure the Token values so they were all values outside the normal ascii range. The keyword tokens are now all < 0x20 to make it easier to find them