Fixed a bug where setting a property on a subclass would fail if the property wasn’t already in the subclass
When searching for a property it could be found in a prototype. But when trying to set that property it needs to be created and set on the subclass. SetProperty could either always add or never add. So I added a 3rd option to AddIfNeeded. When we know the property exists, that option is used.