Changed from pushing 5 prev values on function call to pushing 2
The prevPC and prevObj are now packed into a single value and pushed as PreviousContextA, and prevFrame, prevThis and prevParamCount are packed into a second value and pushed as PreviousContextB. I also added a boolean flag to PreviousContextB which says whether we are returning from a call or constructor. I’ll use this to push the constructed object when needed.