Got ClassTest working
The constructed object was not being returned. So for non-native constructors I added a bool ctor flag to the stack. When a constructor function returns, the previous this pointer (the one used as the this pointer for the constructor function) is pushed onto the stack. But this only happens if the constructor function doesn’t intentionally return a different value, in which case that is used.