Changed TCPEvent firing to be Task based
On the Mac I was using dispatch_async to fire the TCPEvent. This works fine when we’re accessing the socket from native code, but fails when running interpreted code because 2 threads are trying to access the event queue. So now an event schedules a task to run which happens in the same thread as the interpreter. Data from a ReceivedData event is copied into the task and a free list of tasks is kept so multiple events can be in flight at once.