Added a NativeObject class to allow native objects to get properly destroyed
NativeObject is a base class with a virtual destructor. When a PropertyObject is destroyed it looks for a __nativeObject property and if it has one it calls delete on it. Value now holds a NativeObject* rather than a void*. This causes the TCPDelegate to get destroyed when the TCPSocket gets destroyed which properly closes the connected sockets.