Implemented Function.call
You pass the this pointer as the first param and it calls the function with the this pointer and the remainder of the passed params. Allows TCP connection to invoke callback with the proper this value. The TCP test can now open a socket, issue a welcome message and allow text to be typed and returned.