Added LINENO opcode
A new LINENO is generated whenever the line number changes and this value is remembered in the eu. Line number is now included in the error report, so we can see line numbers for runtime errors. Currently the opcode is always generated. But eventually it will be a switch in m8rsim so it can be turned off to avoid the overhead.