Fixed bug in Esp due to odd IdStore issue
There was a line in IdStore.add that caused Esp to crash on boot. The problem is I’m pretty sure that line didn’t run during boot. Compiler error? Anyway, the line had to do with an optimization to IdStore which would make it efficient to add an object right after one was removed. It was rarely used so I just removed it.

I also moved the inline methods using add and remove to be out of line. Because the IdStore functions are inline (because they are templates) it was generating code for every call to Global::addObject and friends. So moving them out of line saved over 1k of flash