Got startup to be more reliable
Sometimes, especially the first time after uploading, the system will hang on startup and the WDT will fire. I’m adding some logic to try to work around that. So far, I’ve added a test to see if everything auto-started. If so, I skip it. But I think what really made things better is adding a print after the got IP callback is made. Maybe it just needs a delay before trying to do MDNS or something?