Fixed crash when a function with upValues is called out of scope.
When a function with upValues is called out of scope, the call stack is not correct for upValue access so we crash. The right solution here is to capture the upValues and create a Closure. But for now, we pass around an inScope flag and remember it across function calls. When we try to access an upValue and this flag isn’t true, we return an empty Value. Right now we determine if a function is in scope by whether or not the call instruction has  constant rcall operand. We should also set it to true if the function has no upValues or if it is a native function