WIP - Implementing Closure capture
When a return value is a function, a Closure is created for it, which captures its up values. To do this I added a Callable base class for both Function and Closure. When a Closure is called it override the load and store upvalue functions and uses the captured values. Not yet working. I’m still sending the Function to startFunction() so the Closure is getting lost.