Switched back to my MDNSResponder.
Turns out the mdns code in the Esp SDK takes a lot of memory, 1.2k of RAM and 1.2k of flash. The original reason got switching is because the UDP calls in the SDK were unreliable. Since then I’ve switched to using lwip and it seems much more stable.