Added syntax and semantics for manual Closure capture
There is now a ‘capture’ keyword. Prefix an expression that results in a Function object and it will convert it into a Closure with all the local variables and this pointer captured. This doesn’t solve Test 5, the JavaScript example where two functions in the same scope can share the same upValue. But for now I’m considering that Closure abuse and only works because of the particular implementation.