Switched from passing ObjectId to Value to passing Object*
ObjectId is still stored in the value. But this makes it easier to get the isFunction() value, which is stored in the Value alongside the id. This is to make it efficient to know if an object is a function without having to get its pointer. This will be used in creating Closures.