WIP - Changing Closures to stay open while in their stack frame and to close when leaving that frame
This breaks Closures for now. Closures are now created whenever their Function is loaded in a register. But they are open and therefore always get their up values from the stack. This makes them work while in their parent function, but not when they leave it. Next step is to keep a list of the open Closures and to close them as needed when the stack frame changes