Fixed bugs in how register are allocated and handled in Parser
There were cases where register allocation was wrong and the register number was getting incremented out of range. So I added an assert for that and found several issues. Most of them had to do with popping a register and then trying to push it back, which didn’t manage the allocated register bookkeeping correctly. So I split the push functions into several versions to keep them all straight. Now I never try to push a register back onto the stack. Instead I use swap to move things around to work on them.