Fixed bug in PRE/POST INC/DEC code generation
These unary ops are different than others. The others simply do the operation on the source reg and then put the result into the dest reg and the source reg is no longer needed. But these must do the same, but then the src reg must be written back to its original location before the register is discarded. I now handle the ops separately and do the proper stack manipulations to generate the right code