Got rid of XPC
It had threading problems and it wasn't doing what it was supposed to do - make it easier to track memory usage.
So now Simulator is a C++ class which is created by Device. It sets up sockets which Device communicates through
and has one custom interface - setFiles, which sets the files known to the simulated device to a filesystem.

This is WIP. GPIO communication is not yet done (need to do a callback?). Also, Simulator also has a bunch of public
API that is no longer needed. And things don't work yet. Seems like they're not getting set up in the right order.