Got m8r running.
This will now be a simulation on Mac. Eventually it will have 3 TCP ports. One for a shell, one for log output, and one for simulated input and output. The first two will be the same on the device version. The third is only on the m8r command line simulation.