Got rid of shared_ptr on Task
Tasks better not go away when they are running in TaskManager. But this is not really a problem since when a Task is destroyed it terminates itself, which takes it out of the TaskManager list. Getting rid of shared_ptr for Task saves about 100 bytes of ram and about 1K of code.