Fixed error handling of files
Now file::open always return a File object. If the open failed it has an error code set. And we should be returning the proper error in all cases. For instance all the test cases are failing because the file system currently doesn't have a "/" file which should be the root directory