Added consoleListener property to Global
To use it you would say 'consoleListener = function(c) { // character came in, it is in 'c' };'

Not hooked up yet. Need to have an internal call to Global. For instance, passing a TCP object would cause Global to redirect print functions from the SystemInterface to this object. And Global would listen for input from this object. When it comes, Global looks for a function in the 'consoleListener' property and calls it with the incoming character. ExecutionUnit::fireEvent probably has to get involved here somehow