Got of passing eu to gc and added registration of eu's
The problem with doing gc for a particular eu is that it does gc for all of memory, not just for that eu. So when you have multiple eu's everything in the one you're not passing gets collected. Now the eu's are registered with Object and it does a gcMark for each eu before doing a collect.