Added a List container class
This is a singly linked list with a Key for sorting. Maybe I should have called it a SortedList or something. The important thing is that it's a singly linked list with one pointer of overhead. So it's less expensive than the std::list class. And since it's sorted it matches perfectly with my needs without going to a std::map which has very high overhead.