Fixed Closure in m8rsh
In order for a closure to hand its result back to its parent, the passed in variable has to be "by reference". Here that mean it must be an object, not a scalar primitive. So I made the 'done' flag into an array with one element and now when the closure set it, the parent see it and the Task exits