Simplified Atoms
It used to be that the built-in shared atoms would be inserted into the general Atom table the first time they were encountered. Then they would be found in the atom table after that. That made all Atoms in the same range. The Atom id was basically just the offset into the Atom string table. Now shared atoms have their enum value as their Atom id. So making a shared atom enum into an Atom is super quick. Non-shared atoms are placed in the table as before. But now their Atom id is offset by 100 to make room for the shared Atom ids. That 100 value is computed in generateSharedAtoms so it will still work if we get more than 100 shared atoms.