Fixed console print
When a TCP connection is opened a lambda is passed to the EU through the Task. This is stored in the EU and made available with a public API. When Global calls any of its print methods it first asks the EU for this consolePrintFunction. If it exists, it is called with a single String param. Otherwise system->printf is called