Got more of Telnet working
Now I'm processing and ignoring incoming Telnet commands. The initial commands sent set the channel in half duplex (no echo) and character at a time. This is as raw as the channel gets. So hitting enter sends CR and that's it. So when sending out data (which only has newlines, I have to insert a CR. And I have to echo characters in Telnet. Currently Telnet is buffering an entire line and then sending it when it received CR. ctl-c is handled and returned immediately. But all other control characters are not handled and are just put in the buffer. I'm also not handling BS or other editing characters. But you can start a telnet connect, have it echo characters correctly and cleanly exit with ctl-c. But opening a second telnet session doesn't work.