Added skeletal exec() function
I'm not sure if I'll go this way. This is intended to act like exec() in C. Pass a file, args and environment and it executes the function. I think it would be better to do this by instantiating a Task, giving it a script to run, loading its environment and telling it to run.