Got rid of ATOM() inline to create internal atoms
Since I changed the Atom store to be the SA enum for internal atoms and the index into the table, offset by 100, for all other, The Program no longer needs to be passed when creating an internal Atom. So I got rid of ATOM() altogether and made it possible to create an internal shared atom with an Atom(SA) constructor. This changes many files, but it's only name changes