Made logical and and or short circuit
Most languages short circuit logical and and or. This means that if the left hand expression evaluates to false when doing a logical and (or to true for a logical or) the right hand side is not evaluated. This saves time when you know evaluating the right hand expression won't change the result. But languages have come to rely on this behavior to avoid things like dereferencing null pointers.

I added logic that would test the left hand result and if the right hand should not be evaluated, I jump over it and instead load a false (for LAND) or true (for LOR). So now I can go:

    if (file && file.valid()) ...

If file is null or not an object it will not call valid() and throw a runtime error.