Got Files closer to properly opening and getting back status
Still issues. The biggest is that FS::open returns a shared_ptr and we need to store a bare File*. I'll probably have to get rid of shared_ptrs.
Changed NativeObject storage to be a proper NativeObject rather than an Object. The value stored in __nativeObject really has to be a
NativeObject. Nothing works right because we're getting valid() on a File* that's been destroyed because of the shared_ptr. And the filename
string passed to FS::open is all wrong.