Changed Value::_value._raw to be a byte array
This value is used to initialize and copy the Value. For Mac it was defined as a uint128_t because it has to be the size of 2 pointers (one to hold the pointer and the other to hold the type and other info. But this causes allocated memory alignment of 8 bytes. So I changed it to an array of bytes that is 2 pointers long. This also gets rid of the need for an ifdef __APPLE__.

The current implementation is probably going to be trouble down the road. Mac requires pointers to be 8 byte aligned. But umm returns 4 byte aligned pointers. But it appears that the addresses ALWAYS end in a 0x04 or 0x0c. As it happens, the pointers in Value come after Type, which is 32 bits. So these pointers always happen to be on an 8 byte alignment. If umm ever returns a pointer that ends in 0x00 or 0x08 I think we're in trouble.