Changed the way memory management is done
I got rid of the linked list. It was creating tons of Mallocators, when you just want one for each MemoryType. Now it is a list of 8 entries: the count and size of each type. Added the tracked allocation to SystemInterface and moved that logic to the non-platform specific code.