Replace all allocated pointers with Mad<...> object
This is an experiment to see if I can make Mallocator return BlockIds rather than pointers. There are several advantages to this:

    1) Makes storage 2 bytes rather than 8 on Mac or 4 on ESP
    2) Makes Mac and ESP behave the same from a storage standpoint
    3) Allows for eventual memory compaction

Others will probably come up. The downside of course is that to convert a Mad to a pointer takes accessing 2 values, a multiply and an add. One of the accesses and the multiply are because I am currently keeping the blockSize as a variable. That's because I'm flexible about the block size. If the available heap is under 256KB then I use 4 byte blocks. Up to 512KB the block size if 8 and for up to 1MB the block size is 16. I can make this a compile time check and then change the pointer conversion to one variable access (the heapBase), one shift by a constant and one add.

The internals are surprisingly hidden. If you used to have a Function* func, you would go func->constants(). Now you have Mad<Function> func and the accessor is the same because Mad overloads ->.

Simple test works, but I'm sure there are pitfalls. For one thing I'm still calling new in a couple of places. But it's a work in progress. I need to find a way to prevent new from being called at all.