Got rid of many more new statements in FS, etc.
I'm passing around Mad values pretty much everywhere now. NativeObject is stored as a Mad, so getNative() now takes a Mad rather than returning a bare pointer.

This is problematic. The Object destructor looks to see if there is a NativeObject and if so destroys it. the problem is it destroys it as a NativeObject, not as the actual subclass. So the size passed to dealloc will be wrong. I need a way to know the actual size of a NativeObject. Maybe a virtual size function?

This breaks the Mac. It can't open the TCP connection because of a FileNotFound error.