WIP - RawMad is in a shambles
Here's what I want. RawMad needs to be a simple (POD) uint16_t so it can be easily stored in arrays. But converting it to a Mad is difficult, partly because of the bool conversion (bools are equivalent to ints). the Id::Raw() concept gets in the way, too. I need to simplify all this