Implemented Fixed memory handling
There are system level calls which need to allocate memory. On ESP this is done with pvPortMalloc/pvPortFree, and on Mac it's done with new/delete. The problem is that these allocations don't know their size and Mallocator needs to know the size to free. So for Fixed memory, I add a 1 block header to the start of the allocation, store the size there, and return a value 1 block ahead of this header. On free I back 1 block, get the size and free that much. Works now for Mac and builds for ESP, but I haven't tested ESP yet