Pass around ROMString rather than const char*
This makes sure that whenever passing around a ROMSTR we pass it as a ROMString so the callee will always process it as such. This doesn't really matter on Mac, but on ESP if you try to read a string that is in Flash as though it's in RAM, it will fail. Still haven't tested ESP, but it should have a better chance of working now