Fixed ROMSTR for ESP
In my previous "fix" I forgot that I put a static string in the ROMSTR inline. Turns out you can't use an inline function at all. Or rather you can, but you won't put the string into Flash. The string will be in RAM and a pointer to it will be in Flash. The whole point of the #define is to store the string literal in a static variable with a storage class of ROMSTR_ATTR. The only way to do that is with the GCC extension of statement expressions. Inside the statement you store the literal in a ROMSTR_ATTR variable and then return it.

The one remaining problem was that storing a pointer to this in a ROMString caused a storage class conflict if the ROMSTR is used in a header file. It's described on the web, but I'm not sure of the root cause. No matter. There was only one instance of this in the code so I moved it to the .cpp file and all is well. ESP uses less RAM than ever because doing this work of consistently putting string literals in Flash caused me to catch a few that I hadn't. RAM usage is down to 30788.