Changed definition of Value
Value used to be 2 pointers. This was very inefficient use since it only ever had to hold either a 64 bit float or a native function pointer or a type and integer or RawMad. But to make it work as an 8 byte value we need to extend the trick used for Float. Since Float is 8 bytes already we had to steal a bit to indicate its type. Now we steal a second bit so we can have up to 3 embedded types. We use 0x01 for Float and 0x02 for NativeFunction. That means we lose another bit of precision in Floats, but we still have 7 decimal digits which is fine. And pointers on both Mac and ESP are at least aligned to 4 bytes, so the lower 2 bits can be used.

Now Value takes up half as much space and memory usage for the test case went from 53KB to 41KB.