Added additional consistency checking
When MEMORY_HEADER is turned on an extra 4 bytes is added to the memory header and it is added to all blocks, allocated and free. It contains a magic value and a type. For all allocated blocks it also stores the block size. This is kept up to date during alloc and free and checked throughout.

The problem is that there is some memory getting trashed but it only happens when MEMORY_HEADER it turned off