Added more checkConsistency callls.
It's now clear that the problem happens when GC walks through the EU stack and marks all the entries as active. Right after this the consistency check fails. It's as though one of the object on the stack is actually deleted or something. But looking at the object itself it looks like _isdestroyed is false.