Found the culprit. Still not sure how to fix it
A Closure gets destroyed, destroying all its UpValues. One of those is object 43277, which gets properly placed in the free list. A bit later MAYBE that same Closure is in the stack. Or is it the same closure? Whatever it is, the same UpValue is in that Closure. The UpValue gets gcMarked, which trashes the nextBlock pointer in object 43277.

Is the Closure destroyed and it's just not marked as such? Or is that UpValue appear in 2 Closures? Is that possible? Right now EU creates UpValues and hands them to the Closure, who is responsible for destroying it. But that is bad ownership. The Closure should create the UpValue and it should be the only owner.