Added some debugging to show all memory allocations
This code has some overhead, but is useful for now. I may remove it or make it more debug friendly. Added a text file with all the allocations done in a 'mem' call.

Biggest problem right now is that the Telnet StateMachine is 1.5K and it gets instantiated 4 times. StateMachine holds state, but maybe the state table can be pulled out into a static?