Added FixedVector class
This class has a size and pointer to vector data. It can be assigned to once and stays that size throughout its lifetime. Aside from the assign() function, which can be called only if size == 0, all other accessors are read-only. This will be used to save space for things like the Instruction vectors, which are never changed once set. The parser will have a normal Instruction Vector which will get filled at compile time and then copied to the FixedVector in the Function.

Since Vector doubles in size when resized, they can be quite a bit larger than need be for fixed arrays of values. For instance, you might have 260 Instructions and it will be in a 512 element Vector, wasting 1KB