Changed OpInfo to be a static array
No more allocating it on the first compile, and no more having that memory lying around allocated forever. Saves about 100 bytes of heap and on ESP about 100 bytes of RAM for a total of 200 bytes. The one downside is that it now does a linear search rather than binary. But there are only 30 elements so it shouldn't be too bad