Added POPX instruction
Often the return value of a function call is discarded. But it still needs to be popped. The POP instruction has the register to pop into as its operand. A new POPX instruction is a single byte which pops the operand and discards it. POP is replaced with POPX when the Parser sees the the result is being discarded.

Saves 20 bytes in the test