WIP - Experimenting with calling Functions directly.
Currently every time a function is created and store (either a Function statement or expression) a CLOSURE opcode is generated. That makes every call a closure call, which is inefficient when you don't need it. A closure should be created later, when you return a Function or maybe when you assign it. This patch shuts off all Closure generation code. With the simple test example here it works fine. It certainly won't when you need an actual closure. But I also need to test access to variables outside the scope of the function, which is another use for the Closure upvalues.