Fixed Closures
Everything working again. Now a Closure is only created when it is pushed on the stack. There will probably be other cases, e.g., when a function is assigned to a variable or something.

Now functions are usually called directly. If they have an upValue, it is necessarily an open upValue so the value is accessed directly from the stack. This saves 340 bytes from having to create fewer Closures and fewer closed upValues.