Split MaterArray out of MaterObject
It's nice to have a really generic and efficient Object class. Where you can efficiently store both numeric and string keys. But it's expensive carrying around a Map and a Vector in every class. So I made a MaterArray class. It accepts only numeric keys and stores values in a Vector. MaterObject now stores everything in a Map. You can still do obj[10] = 5; in MaterObject. But it will store the number 5 in a property named "10". And you can still access properties either with obj.foo or obj["foo"]. But MaterArray can only take numeric values. If you try to get or set properties it will fail (unless it's the built-in length property).

Saves another 100 bytes