Some ESP Success!
I've been struggling with many toolchains and being unsuccessful with all but Arduino. But the Arduino IDE is very restrictive when it comes to the build and I don't want to restructure the entire system just to accommodate it.

Then I stumbled on the ESP IDF RTOS SDK. I have avoided it until now because I thought it was too big. But it makes much better use of RAM, so it can use some areas that the NONOS SDK doesn't. So far with a small test it makes 100K of memory available, which is about double what was available with the NONOS or Arduino. And it has a decent build system with Makefiles. So I've started using it. So far I have a small startup main which just does some printing. But it's a start