Fixed ESP crash
Turns out it really was running out of stack space. SLRE (which supposedly stands for "super lightweight" regex library (although it's also the author's initials) turns out to be no so lightweight after all. Internal arrays add up to 2600 bytes. And with a 3500 byte stack that doesn't leave much. As it turns out the needs for the formatter are small so I was able to reduce that to just 90 bytes. I also increased the stack to 5K just to be safe. And now it all works