WIP - Get RTOS to put more things in Flash
RTOS automatically puts some things in Flash. Strings go there, although I'm not quite sure if special accessors are needed. And arrays that are POD go in there. For instance the dispatchTable in ExecutionUnit goes into Flash, but the SharedAtoms do not. So I'm going through to see if I can change more things to POD. So far Task was crashing trying to put its StaticProperties into Flash because StaticProperty isn't POD. So I added a StaticFunctionProperty that is POD. Now I need to convert all the other StaticProperties to use it