Changed how atom table is stored
Atom table used to be stored as a negative size byte followed by the chars of the atom. This was done to avoid putting embedded NUL chars in the table, so strstr could be used. The problem with that is that you can't return a pointer to an embedded string because there's no ending NUL. To allow a proper string pointer to be returned and to not have to copy the atom into a String, I changed the table to be a series of NUL terminated strings and wrote a custom strstr like function that can deal with embedded NULs.

Now I can change stringFromAtom to return a const char*