Added toStringPointer method
Sometimes you know that a Value can be converted to a string without constructing a String object. This is true for Values that are Strings, StringLiterals and Ids, as well as Null and undefined. In these cases we can return a const char* and avoid the cost of creating a String object. Saves a bit more code space