WIP - Mostly working
Got rid of Instruction. Parser::addCode now takes Op, up to 3 regs and a number. Based on OpInfo the proper instruction is written out. Handles the special case of using RETI when less than 4 return values. LINENO is emitted as needed. Before instruction for all bure JF, JT and JMP and after for those 3. Never output if the instruction is LINENO of course.

hello.m8r and mrsh run correctly. mem runs but gets a runtime error at the end