Fixed bug in lfs_file_config
I use this struct now so I can pass a buffer size to LFS file operations. This keeps LFS from allocating any memory. But when I did it I made it a local variable and it needs to have the lifetime of the open file. So I made it an instance variable and now it no longer crashes