WIP - Implemented Timer functionality.
This will be built on top of Task which already has the concept of executing the task at some time in the future. Implemented the repeating functionality, but there is no constructor yet that takes the signature Timer(Duration, Behavior {Once, Repeating}, function).

Crashes when running hello.m8r, even without any changes to that code