Got rid of SystemInterface::currentMicroseconds
The ESP RTOS SDK fully supports all the <ctime> and <chrono> functions so now SystemTime::now() uses them directly to get the current time.

Note that currently time on the ESP is not getting set. It apparently has an accurate RTC running, but it gets reset to 0 on reboot. In the lwip library this is an SNTP library. You can start it running and it will get the current time and then keep it updated. This requires getting WiFi up and running of course.