WIP - Restructuring how TaskManager works
Making so Tasks are always in TaskManager list. Each one has a state. Tasks in the ready list are run one at a time and then added to the end of the list. Tasks waiting for a delay or event are in a separate list and are placed in the ready list when ready. Need to figure how how to deal with event that comes in while a Task is waiting for a delay.